requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
bcrypt==4.0.1
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
dateparser==1.1.7
defusedxml==0.7.1
Django==4.1.7
django-appconf==1.0.5
django-environ==0.10.0
django-grappelli==3.0.4
django-imagekit==4.1.0
django-semantic-admin==0.3.8
django-summernote==0.8.20.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.2
idna==3.4
oauthlib==3.2.2
paramiko==2.8.1
pilkit==2.0
Pillow==9.4.0
promise==2.3
prompt-toolkit==2.0.9
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
Rx==1.6.3
six==1.16.0
social-auth-core==4.3.0
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
sshtunnel==0.4.0
tornado==5.1.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
gunicorn==20.1.0